working

+++ app.js
@@ -4,7 +4,6 @@ const express = require('express');
 const mongoose = require('mongoose');
 const cors = require('cors');
 
-
 const app = express();
 
 // app.use(express.urlencoded({ extended: true }));
@@ -13,15 +12,14 @@ app.use(express.static(__dirname + '/public'));
 
 app.use(cors());
 
-const port = process.env.PORT || 3000
+const port = process.env.PORT || 3000;
 
 // const url = process.env.MONGO_URL;
 const url = 'mongodb+srv://dbUser:' + process.env.MONGO_PW + '@cluster0.dq3y7.mongodb.net/' + process.env.MONGO_DB;
 
-mongoose.connect(url, { useNewUrlParser: true, useUnifiedTopology: true }).catch(err => {
+mongoose.connect(url, { useNewUrlParser: true, useUnifiedTopology: true }).catch((err) => {
   console.error('😱Something went wrong: ' + err);
 });
-
 
 // ==========================================================================
 
@@ -40,8 +38,8 @@ app.get('/', (req, res) => {
     <h3>/users/:userID</h3>
     <p>GET - PUT - PATCH - DELETE</>
     `
-    );
-});  
+  );
+});
 
 // ==========================================================================
 // Blogposts
@@ -60,7 +58,7 @@ const blogpostSchema = mongoose.Schema({
   image2URL: String,
   wiki: String,
   utcOffset: Number,
-  authorID: String
+  authorID: String,
 });
 
 const Blogpost = mongoose.model('Blogpost', blogpostSchema);
@@ -81,11 +79,11 @@ app
   })
   .post((req, res) => {
     console.log('POST new blogpost');
- 
+
     const newBlogpost = new Blogpost(req.body);
     newBlogpost.save((err) => {
       if (!err) {
-        res.send('Succesfully added a new blogpost: ' + newBlogpost.name );
+        res.send('Succesfully added a new blogpost: ' + newBlogpost.name);
       } else {
         res.send(err);
       }
@@ -103,60 +101,49 @@ app
   });
 
 ///////////////////// request for a specific blogpost
-app.route('/blogposts/:blogpostID')
+app
+  .route('/blogposts/:blogpostID')
   .get((req, res) => {
     console.log('GET one blogpost');
-    Blogpost.findOne(
-      {_id: req.params.blogpostID}, (err, foundBlogpost) => {
-        if(foundBlogpost) {  
-          res.send(foundBlogpost);
-        } else {
-          res.send("No Blogpost found");
-        }
+    Blogpost.findOne({ _id: req.params.blogpostID }, (err, foundBlogpost) => {
+      if (foundBlogpost) {
+        res.send(foundBlogpost);
+      } else {
+        res.send('No Blogpost found');
       }
-    )
+    });
   })
 
   .put((req, res) => {
     console.log('PUT');
     // console.log(req.body);
-    Blogpost.replaceOne(
-      {_id: req.params.blogpostID},
-      req.body,
-      (err, results) => {
-        if(!err) {
-          res.send(results);
-        } else {
-          res.send(err)
-        }
+    Blogpost.replaceOne({ _id: req.params.blogpostID }, req.body, (err, results) => {
+      if (!err) {
+        res.send(results);
+      } else {
+        res.send(err);
       }
-    )
+    });
   })
 
-  .patch((req,res) => {
+  .patch((req, res) => {
     console.log('PATCH');
-    Blogpost.updateOne(
-      {_id: req.params.blogpostID},
-      {$set: req.body},
-      (err, results) => {
-        if(!err) {
-          res.send(results);
-        } else {
-          res.send(err)
-        }
+    Blogpost.updateOne({ _id: req.params.blogpostID }, { $set: req.body }, (err, results) => {
+      if (!err) {
+        res.send(results);
+      } else {
+        res.send(err);
       }
-    )
+    });
   })
   .delete((req, res) => {
-    Blogpost.deleteOne(
-      {_id: req.params.blogpostID},
-      (err) => {
-        if(!err) {
-          res.send("🚫successfully deleted ONE Blogpost");
-        } else {
-          res.send(err)
-        }
-      })
+    Blogpost.deleteOne({ _id: req.params.blogpostID }, (err) => {
+      if (!err) {
+        res.send('🚫successfully deleted ONE Blogpost');
+      } else {
+        res.send(err);
+      }
+    });
   });
 
 // ==========================================================================
@@ -167,7 +154,7 @@ const userSchema = mongoose.Schema({
   name: String,
   email: String,
   password: String,
-  profilePic: String
+  profilePic: String,
 });
 
 const User = mongoose.model('User', userSchema);
@@ -187,11 +174,11 @@ app
   })
   .post((req, res) => {
     console.log('POST new user', req.body);
- 
+
     const newUser = new User(req.body);
     newUser.save((err) => {
       if (!err) {
-        res.send('Succesfully added a new user: ' + newUser.name );
+        res.send('Succesfully added a new user: ' + newUser.name);
       } else {
         res.send(err);
       }
@@ -208,63 +195,52 @@ app
     });
   });
 ///////////////////// request for a specific user
-app.route('/users/:userID')
+app
+  .route('/users/:userID')
   .get((req, res) => {
     console.log('GET one user');
-    User.findOne(
-      {_id: req.params.userID}, (err, foundUser) => {
-        if(foundUser) {  
-          res.send(foundUser);
-        } else {
-          res.send("No User found");
-        }
+    User.findOne({ _id: req.params.userID }, (err, foundUser) => {
+      if (foundUser) {
+        res.send(foundUser);
+      } else {
+        res.send('No User found');
       }
-    )
+    });
   })
 
   .put((req, res) => {
     console.log('PUT user');
     // console.log(req.body);
-    User.replaceOne(
-      {_id: req.params.userID},
-      req.body,
-      (err, results) => {
-        if(!err) {
-          res.send(results);
-        } else {
-          res.send(err)
-        }
+    User.replaceOne({ _id: req.params.userID }, req.body, (err, results) => {
+      if (!err) {
+        res.send(results);
+      } else {
+        res.send(err);
       }
-    )
+    });
   })
 
-  .patch((req,res) => {
+  .patch((req, res) => {
     console.log('PATCH user');
-    User.updateOne(
-      {_id: req.params.userID},
-      {$set: req.body},
-      (err, results) => {
-        if(!err) {
-          res.send(results);
-        } else {
-          res.send(err)
-        }
+    User.updateOne({ _id: req.params.userID }, { $set: req.body }, (err, results) => {
+      if (!err) {
+        res.send(results);
+      } else {
+        res.send(err);
       }
-    )
+    });
   })
   .delete((req, res) => {
-    User.deleteOne(
-      {_id: req.params.userID},
-      (err) => {
-        if(!err) {
-          res.send("🚫successfully deleted ONE User");
-        } else {
-          res.send(err)
-        }
-      })
+    User.deleteOne({ _id: req.params.userID }, (err) => {
+      if (!err) {
+        res.send('🚫successfully deleted ONE User');
+      } else {
+        res.send(err);
+      }
+    });
   });
 
-  // ==========================================================================
+// ==========================================================================
 
 app.listen(port, () => {
   console.log('=======================================================');
